+++ CHANGELOG.md
@@ -4,7 +4,19 @@ All notable changes to this project will be documented in this file.
 The format based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Initial Version] - 2020-04-12
+## [0.0.2] 2020-04-13
+
+### Added
+
+- BadRequest Error
+- RecordNotFound Error
+- NotFound Error
+
+### Changed
+
+- Default CustomError. Added `this.errorType` and `this.detail` for additional debugging information.
+
+## [0.0.1] - 2020-04-12 - Initial Version
 
 ### Added